+++ packages/ui/src/screens/jobs/Filters/index.tsx
@@ -20,6 +20,7 @@ import DataSearch from './DataSearch';
 import DataSearchTip from './DataSearch/Tip';
 import JobNameFilter from './JobName';
 import FilterPresetsMenu from './Presets';
+import FilterBuilder from './Builder';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -146,6 +147,7 @@ export default function JobsFilters() {
           <MenuItem value={OrderEnum.Asc}>ASC</MenuItem>
         </TextField>
         <FilterPresetsMenu className={cls.clearButton} />
+        <FilterBuilder className={cls.clearButton} />
         <JobNameFilter className={cls.nameField} />
         <DataSearch className={cls.dataSearchField} />
         {hasActiveFilters && (